Add handling of non-equal forecast lengths.
The policy logic introduced here for accessors is not good, and I will
look into it further, since this only works if the accessor and the
time series have 'comparable' time axes, i.e. have the same points up
till the point where the accessor continues but the time series don't.